+++ internal/api/api.go
@@ -39,6 +39,7 @@ const (
 	ParamOrderBy    = "orderby"
 	ParamPrecision  = "precision"
 	ParamProperties = "properties"
+	ParamSortBy     = "sortby"
 	ParamTransform  = "transform"
 
 	OrderByDirSep = ":"
@@ -92,6 +93,7 @@ var ParamReservedNames = []string{
 	ParamOrderBy,
 	ParamPrecision,
 	ParamProperties,
+	ParamSortBy,
 	ParamTransform,
 }
 
@@ -231,7 +233,7 @@ type RequestParam struct {
 	Bbox          *data.Extent
 	Properties    []string
 	GroupBy       []string
-	OrderBy       []data.Ordering
+	SortBy        []data.Sorting
 	Precision     int
 	TransformFuns []data.TransformFunction
 	Values        NameValMap
@@ -456,9 +458,17 @@ var ConformanceSchema openapi3.Schema = openapi3.Schema{
 
 var conformance = Conformance{
 	ConformsTo: []string{
-		"http://www.opengis.net/spec/wfs-1/3.0/req/core",
-		"http://www.opengis.net/spec/wfs-1/3.0/req/geojson",
-		"http://www.opengis.net/spec/wfs-1/3.0/req/html",
+		"http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/core",
+		"http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/oas3",
+		"http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/geojson",
+		"http://www.opengis.net/spec/ogcapi-features-1/1.0/conf/html",
+		"http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/core",
+		"http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/landing-page",
+		"http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/json",
+		"http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/html",
+		"http://www.opengis.net/spec/ogcapi-common-1/1.0/conf/oas30",
+		"http://www.opengis.net/spec/ogcapi-common-2/1.0/conf/collections",
+		"http://www.opengis.net/spec/ogcapi-common-2/1.0/conf/simple-query",
 	},
 }