+++ internal/api/openapi.go
@@ -143,6 +143,16 @@ func GetOpenAPIContent(urlBase string) *openapi3.Swagger {
 			AllowEmptyValue: false,
 		},
 	}
+	paramSortBy := openapi3.ParameterRef{
+		Value: &openapi3.Parameter{
+			Name:            "sortby",
+			Description:     "Column to sort by.",
+			In:              "query",
+			Required:        false,
+			Schema:          &openapi3.SchemaRef{Value: openapi3.NewStringSchema()},
+			AllowEmptyValue: false,
+		},
+	}
 	return &openapi3.Swagger{
 		OpenAPI: "3.0.0",
 		Info: openapi3.Info{
@@ -252,6 +262,7 @@ func GetOpenAPIContent(urlBase string) *openapi3.Swagger {
 						&paramOffset,
 						&paramBbox,
 						&paramProperties,
+						&paramSortBy,
 						&paramTransform,
 						/* TODO
 						&openapi3.ParameterRef{
@@ -369,7 +380,9 @@ func GetOpenAPIContent(urlBase string) *openapi3.Swagger {
 						&paramOffset,
 						&paramBbox,
 						&paramProperties,
+						&paramSortBy,
 						&paramTransform,
+
 						/* TODO
 						&openapi3.ParameterRef{
 							Value: &openapi3.Parameter{