+++ internal/data/catalog.go
@@ -21,6 +21,8 @@ import (
 
 // Catalog tbd
 type Catalog interface {
+	SetIncludeExclude(includeList []string, excludeList []string)
+
 	Tables() ([]*Table, error)
 
 	// TableByName returns the table with given name.
@@ -57,7 +59,7 @@ type TransformFunction struct {
 	Arg  []string
 }
 
-type Ordering struct {
+type Sorting struct {
 	Name   string
 	IsDesc bool // false = ASC (default), true = DESC
 }
@@ -76,7 +78,7 @@ type QueryParam struct {
 	// Columns is the list of columns to return
 	Columns       []string
 	GroupBy       []string
-	OrderBy       []Ordering
+	SortBy        []Sorting
 	Precision     int
 	TransformFuns []TransformFunction
 }